fix(typescript): resolve all remaining TypeScript errors from strictest mode migration
Systematically fixed 29 TypeScript errors across test files after implementing
@tsconfig/strictest configuration. Key changes:

- Applied AcceleratePromise compatibility patterns for Prisma mocks
- Fixed type assertions in mock infrastructure (mockContext.ts, serviceHelpers.ts)
- Cleaned up duplicate pinballmapService test files (renamed V2 → main)
- Added proper type casting for service and database mocks
- Fixed ESLint unbound-method violations in notification tests
- Updated Betterer baseline to reflect zero remaining errors

All validation passes: TypeScript (0 errors), ESLint (0 problems),
tests (22/22 passing), build successful.
