+++ src/test/helpers/serviceHelpers.ts
@@ -6,9 +6,11 @@ export function mockServiceMethod<T extends keyof ServiceFactory>(
   serviceName: T,
   methodName: keyof ReturnType<ServiceFactory[T]>,
   implementation?: (...args: unknown[]) => unknown,
-): jest.MockedFunction<unknown> {
-  const service = services[serviceName]() as Record<string, unknown>;
-  const method = service[methodName] as jest.MockedFunction<unknown>;
+): jest.MockedFunction<(...args: unknown[]) => unknown> {
+  const service = services[serviceName]() as unknown as Record<string, unknown>;
+  const method = service[methodName as string] as jest.MockedFunction<
+    (...args: unknown[]) => unknown
+  >;
   if (implementation) {
     method.mockImplementation(implementation);
   }
@@ -22,11 +24,18 @@ export function resetAllServiceMocks(
   for (const key in services) {
     if (typeof services[key as keyof ServiceFactory] === "function") {
       const service = (
-        services[key as keyof ServiceFactory] as () => Record<string, unknown>
+        services[key as keyof ServiceFactory] as unknown as () => Record<
+          string,
+          unknown
+        >
       )();
       for (const method in service) {
         if (jest.isMockFunction(service[method])) {
-          (service[method] as jest.MockedFunction<unknown>).mockClear();
+          (
+            service[method] as unknown as jest.MockedFunction<
+              (...args: unknown[]) => unknown
+            >
+          ).mockClear();
         }
       }
     }
